// app/src/engine/types/Ptz.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PtzProtocol } from './PtzProtocol';

export type Ptz = {
  /**
   * The camera's IP address or name.
   */
  host: string;
  /**
   * 0: the usual port for the protocol.
   */
  port: number;
  protocol: PtzProtocol;
  /**
   * Names for the preset buttons (empty: just the number).
   */
  presets: Array<string>;
};